Update new cody web UI (side panel and standalone chat  page)
