+++ client/web/src/cody/chat/new-chat/components/chat-history-list/ChatHistoryList.module.scss
@@ -5,6 +5,12 @@
     list-style: none;
     margin: 0;
     padding: 0;
+    position: relative;
+    height: 100%;
+
+    li {
+        margin-right: 0.5rem;
+    }
 }
 
 .history-item {
@@ -13,11 +19,11 @@
     flex-direction: column;
     margin-bottom: 0;
     padding: 0.5rem 0.75rem;
-    border: none;
     cursor: pointer;
     border-radius: 0.25rem;
     text-align: left;
     background: var(--body-bg);
+    border: 1px solid var(--border-color);
 
     &:hover {
         background: var(--color-bg-1);
@@ -36,4 +42,13 @@
         color: var(--icon-color);
         visibility: hidden;
     }
+}
+
+.footer {
+    position: sticky;
+    bottom: 0;
+    margin-top: auto;
+    padding: 0.75rem 0.5rem 0.75rem 0;
+    background: var(--body-bg);
+    border-top: 1px solid var(--border-color);
 }
