+++ client/web/src/cody/chat/new-chat/components/chat-ui/ChatUI.module.scss
@@ -1,10 +1,13 @@
 .chat {
     --vscode-editor-background: var(--body-bg);
     --vscode-editor-foreground: var(--body-color);
-    --vscode-input-background: rgb(255 255 255);
+    --vscode-input-background: var(--search-box-color);
     --vscode-input-foreground: var(--body-color);
     --vscode-textLink-foreground: var(--primary);
     --vscode-input-border: var(--border-color-2);
+    --vscode-inputOption-activeBackground: var(--search-input-token-filter);
+    --vscode-inputOption-activeForeground: var(--body-color);
+    --mention-color-opacity: 100%;
 
     h3 {
         font-size: inherit;
@@ -13,6 +16,10 @@
 
     ul {
         margin: 0;
+    }
+
+    a {
+        color: var(--link-color) !important;
     }
 }
