+++ client/web/src/cody/chat/new-chat/components/chat-ui/ChatUi.tsx
@@ -1,8 +1,9 @@
 import type { FC } from 'react'
 
 import classnames from 'classnames'
+import { CodyWebChat } from 'cody-web-experimental'
 
-import { CodyWebChat } from '@sourcegraph/cody-web'
+import 'cody-web-experimental/dist/style.css'
 
 import styles from './ChatUI.module.scss'
