+++ client/web/src/cody/sidebar/new-cody-sidebar/NewCodySidebar.module.scss
@@ -1,0 +1,17 @@
+.root {
+    display: flex;
+    flex-direction: column;
+    height: 100%;
+    overflow: hidden;
+    filter: drop-shadow(0 0 4px rgba(36, 41, 54, 0.1));
+}
+
+.header {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 0.5rem 1rem;
+    font-size: 0.875rem;
+    background-color: var(--input-bg);
+    border-bottom: 1px solid var(--border-color-2);
+}
